requirements.txt
acme==1.21.0
aenum==3.1.15
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
amqp==5.2.0
annotated-types==0.6.0
anyio==4.3.0
anyjson==0.3.3
appdirs==1.4.4
asgiref==3.8.1
async-timeout==4.0.3
attrs==21.2.0
autobahn==21.11.1
Automat==20.2.0
Babel==2.8.0
base58==1.0.3
bcrypt==3.2.0
beautifulsoup4==4.12.3
billiard==4.2.0
blinker==1.4
Brotli==1.1.0
cached-property==1.5.2
cachetools==5.3.3
cashfree-pg==4.1.3
cbor==1.0.0
celery==5.4.0
certbot==1.21.0
certbot-nginx==1.21.0
certifi==2020.6.20
cffi==1.15.0
channels==4.1.0
channels-redis==4.2.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.2.0
cloud-init==24.1.3
colorama==0.4.4
command-not-found==0.3
ConfigArgParse==1.5.3
configobj==5.0.6
constantly==15.1.0
contourpy==1.2.1
cryptography==3.4.8
cycler==0.12.1
Cython==0.29.28
daphne==3.0.2
dataclasses-json==0.6.6
dbus-python==1.2.18
defusedxml==0.8.0rc2
diff-match-patch==20230430
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.0.6
django-allauth==0.62.1
django-celery-results==2.5.1
django-channels==0.7.0
django-ckeditor==6.7.1
django-cors-headers==4.3.1
django-import-export==4.0.2
django-jazzmin==3.0.0
django-js-asset==2.2.0
django-taggit==5.0.1
django-unfold==0.22.0
djangorestframework==3.15.1
dnspython==2.6.1
docopt==0.6.2
ec2-hibinit-agent==1.0.0
ecdsa==0.18.0b1
email_validator==2.1.1
ephem==4.1.3
exceptiongroup==1.2.1
faiss-gpu==1.7.2
fastapi==0.111.0
fastapi-cli==0.0.3
ffmpy==0.3.2
filelock==3.14.0
fireworks-ai==0.14.0
flatbuffers===1.12.1-git20200711.33e2d80-dfsg1-0.6
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
google==3.0.0
google-ai-generativelanguage==0.6.3
google-api-core==2.19.0
google-api-python-client==2.129.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud==0.34.0
google-generativeai==0.5.3
googleapis-common-protos==1.63.0
gradio==4.31.4
gradio_client==0.16.4
greenlet==3.0.3
groq==0.5.0
grpcio==1.63.0
grpcio-status==1.62.2
gunicorn==22.0.0
gyp==0.1
h11==0.14.0
hibagent==1.0.1
httpcore==1.0.5
httplib2==0.20.2
httptools==0.6.1
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.23.0
hyperlink==21.0.0
idna==3.3
ijson==3.2.3
importlib-metadata==4.6.4
importlib_resources==6.4.0
incremental==21.3.0
jeepney==0.7.1
Jinja2==3.0.3
josepy==1.10.0
jsonpatch==1.33
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
kiwisolver==1.4.5
kombu==5.3.7
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
langchain-fireworks==0.1.2
langchain-groq==0.1.3
langchain-text-splitters==0.0.1
langsmith==0.1.57
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lz4==3.1.3+dfsg
markdown-it-py==3.0.0
MarkupSafe==2.0.1
marshmallow==3.21.2
matplotlib==3.9.0
mdurl==0.1.2
mnemonic==0.19
more-itertools==8.10.0
msgpack==1.0.8
multidict==6.0.5
mypy-extensions==1.0.0
netifaces==0.11.0
numpy==1.26.4
oauthlib==3.2.0
openai==1.29.0
orjson==3.10.3
packaging==23.2
pandas==2.2.2
parsedatetime==2.6
passlib==1.7.4
peewee==3.17.5
pexpect==4.8.0
pillow==10.3.0
Pint==0.23
platformdirs==4.2.1
ply==3.11
prettytable==3.10.0
prompt-toolkit==3.0.28
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
py-ubjson==0.16.1
pyasn1==0.4.8
pyasn1-modules==0.2.1
pybcj==1.0.2
pycasreg==0.1.0
pycparser==2.21
pycryptodomex==3.20.0
pydantic==1.10.15
pydantic_core==2.18.2
pydub==0.25.1
Pygments==2.18.0
PyGObject==3.42.1
PyHamcrest==2.0.2
PyICU==2.8.1
PyJWT==2.8.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pypng==0.0.20
pyppmd==1.1.0
PyPrind==2.11.3
PyQRCode==1.2.1
pyRFC3339==1.1
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu3
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
python-json-logger==2.0.7
python-magic==0.4.24
python-memcached==1.58
python-multipart==0.0.9
python-snappy==0.5.3
python3-openid==3.2.0
PyTrie==0.4.0
pytz==2022.1
pyxlsb==1.0.10
PyYAML==5.4.1
pyzstd==0.15.10
redis==5.0.4
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==0.9.1
rfc3986==2.0.0
rich==13.7.1
rsa==4.9
ruff==0.4.4
SecretStorage==3.3.1
semantic-version==2.10.0
sentry-sdk==1.32.0
service-identity==18.1.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sortedcontainers==2.1.0
sos==4.5.6
soupsieve==2.5
SQLAlchemy==2.0.30
sqlparse==0.5.0
ssh-import-id==5.11
starlette==0.37.2
systemd-python==234
tablib==3.5.0
tenacity==8.3.0
texttable==1.7.0
tomlkit==0.12.0
toolz==0.12.1
tqdm==4.66.4
Twisted==22.1.0
txaio==21.2.1
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
u-msgpack-python==2.3.0
ubuntu-pro-client==8001
ufw==0.36.1
ujson==5.10.0
unattended-upgrades==0.1
unicodecsv==0.14.1
Unidecode==1.3.8
uritemplate==4.1.1
urllib3==2.0.7
uvicorn==0.29.0
uvloop==0.19.0
vine==5.1.0
voluptuous==0.14.2
wadllib==1.3.6
watchfiles==0.21.0
wcwidth==0.2.5
websockets==11.0.3
Whoosh==2.7.4
wrapt==1.16.0
wsaccel==0.6.3
xlrd==2.0.1
XlsxWriter==3.2.0
yarl==1.9.4
zipp==1.0.0
zope.component==4.3.0
zope.event==4.4
zope.hookable==5.1.0
zope.interface==5.4.0
